Allow to tag feeds

+++ um/models.py
@@ -7,7 +7,7 @@ class UMProfile(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE,
                                 related_name='um_profile')
     night_mode = models.BooleanField(default=False)
-    items_per_page = models.PositiveIntegerField(
+    items_per_page = models.PositiveSmallIntegerField(
         default=20,
         validators=[MinValueValidator(1), MaxValueValidator(200)]
     )
